+++ src/win32/main.c
@@ -163,7 +163,7 @@ int wWinMain(HINSTANCE instance,
     r_list.command_list_size = MB(16),
     r_list.command_list_base = m_alloc_nozero(&win32_arena, r_list.command_list_size, 16);
     r_list.max_immediate_icount = 1 << 20;
-    r_list.immediate_indices = m_alloc_array_nozero(&win32_arena, r_list.max_immediate_icount, uint16_t);
+    r_list.immediate_indices = m_alloc_array_nozero(&win32_arena, r_list.max_immediate_icount, uint32_t);
     r_list.max_immediate_vcount = 1 << 19;
     r_list.immediate_vertices = m_alloc_array_nozero(&win32_arena, r_list.max_immediate_vcount, vertex_immediate_t);
     r_set_command_list(&r_list);
